Increases Efficiency by Skipping Creation of List
Avoids creating a list and then iterating through that list. Instead, we iterate through range(2, n) and create new_n which is the sum of n1 + n2, and update n2 to be new_n and n1 to be n2. This way, the fibonacci numbers are updated but no initial list must be stored to be iterated through. The values of n1 and n2 are just updated.